Refactor: removing marklogic-spark-langchain4j
The langchain4j-specific stuff was moved to marklogic-langchain4j, and the non-specific stuff was sent to marklogic-spark-api.

Next up - moving all the tests into a separate subproject that can depend on Java 17.
