--- marklogic-spark-api/README.md
@@ -1,2 +1,0 @@
-This module defines the public API that marklogic-spark-langchain4j can depend on without depending on Spark or the
-connector itself.
