+++ marklogic-spark-connector/build.gradle
@@ -52,7 +52,7 @@ dependencies {
 
   // Once we update langchain4j to 0.36.0 or higher, which requires Java 17, this will need to become a
   // testImplementation dependency, and we'll figure out how to include them separately in Flux.
-  shadowDependencies (project(":marklogic-spark-langchain4j")) {
+  shadowDependencies (project(":marklogic-langchain4j")) {
     exclude group: "com.fasterxml.jackson.core"
     exclude group: "com.fasterxml.jackson.dataformat"
   }
@@ -61,12 +61,10 @@ dependencies {
   // include the necessary parsers.
   shadowDependencies "org.apache.tika:tika-core:3.0.0"
 
-  // Oddly need this specified here too, otherwise compileTestJava fails. Not sure why it's not coming via the
-  // "marklogic-spark-langchain4j" dependency.
-  testImplementation (project(":marklogic-langchain4j")) {
-    exclude group: "com.fasterxml.jackson.core"
-    exclude group: "com.fasterxml.jackson.dataformat"
-  }
+//  testImplementation (project(":marklogic-langchain4j")) {
+//    exclude group: "com.fasterxml.jackson.core"
+//    exclude group: "com.fasterxml.jackson.dataformat"
+//  }
 
   // Supports testing the embedder feature.
   testImplementation "dev.langchain4j:langchain4j-embeddings-all-minilm-l6-v2:0.35.0"
