--- marklogic-spark-langchain4j/build.gradle
@@ -1,19 +1,0 @@
-/**
- * This module acts as a bridge between the Spark connector and the marklogic-langchain4j module, which we may
- * eventually make available separately from the connector.
- */
-dependencies {
-  compileOnly "org.apache.spark:spark-sql_2.12:" + sparkVersion
-  compileOnly "com.marklogic:marklogic-client-api:7.1.0"
-
-  implementation project(":marklogic-spark-api")
-
-  // API dependency so that public class in marklogic-langchain4j - specifically Util - can be accessed
-  // in the connector.
-  api (project(":marklogic-langchain4j")) {
-    // These need to be excluded here so they don't sneak in when this module is depended on by the
-    // marklogic-spark-connector module.
-    exclude group: "com.fasterxml.jackson.core"
-    exclude group: "com.fasterxml.jackson.dataformat"
-  }
-}
