+++ settings.gradle
@@ -1,6 +1,5 @@
-include "marklogic-langchain4j"
 include "marklogic-spark-api"
-include "marklogic-spark-langchain4j"
+include "marklogic-langchain4j"
 include "marklogic-spark-connector"
 include "test-app"
 include "code-coverage-report"
